% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridge.linearmap.R
\name{bridge.linearmap}
\alias{bridge.linearmap}
\title{Bridging Two Ideal Point Estimates with Linear Transformation Method}
\usage{
bridge.linearmap(
  ip1,
  ip2,
  anchorrows.ip1,
  anchorrows.ip2,
  method = "procrustes",
  trans.ip2 = TRUE,
  opt = FALSE,
  opt.iter.n = 10000,
  opt.sample.n = 30,
  opt.th.inline = 0.5,
  blend = TRUE,
  blend.th1 = 0.05,
  blend.th2 = 0.15
)
}
\arguments{
\item{ip1}{Matrix or data.frame of 'reference' ideal points 
(i.e., \code{ip2} ideal points will be transformed and mapped to \code{ip1} space).
Rows are resepondents and columns are ideal point dimensions. 
The current code only allows 2 dimensions.}

\item{ip2}{Matrix or data.frame of ideal points to be transformed.}

\item{anchorrows.ip1}{Vector of row number of anchoring respondents in \code{ip1}.}

\item{anchorrows.ip2}{Vector of row number of anchoring respondents in \code{ip2}.
Must be the same length as \code{anchorrows.ip1}.}

\item{method}{Method of bridging. Currently, following methods are aviailable:
\itemize{
  \item \code{"procrustes"} (default): Procrustes transformation method. 
  Based on anchor cases, this method provides restricted non-parametric procedure to 
  find optimal transformation matrix to bridge ideal point estimates.
  \item \code{"homography"}: Homography transformation method. 
  Based on anchor cases, this method provides non-parametric procedure to 
  find optimal transformation matrix to bridge ideal point estimates.
  \item \code{"olsmap"}: OLS mapping method, Based on anchor cases, 
  use OLS regression to map \code{d2} ideal point coordinates on 
  \code{d1} ideal point space.  
}}

\item{trans.ip2}{If \code{TRUE} (default), transform \code{ip2} to map them on \code{ip1} space.
If \code{FALSE}, transform \code{ip1} to map them on \code{ip2} space.}

\item{opt}{If \code{TRUE}, conduct optimization of transformation through RANSAC
(random sample consensus). The default is \code{FALSE}.}

\item{opt.iter.n}{Number of iteration in the optimization of transformation 
matrix.}

\item{opt.sample.n}{Size of anchoring respondents to be sub-sampled at each iteration of 
optimization.}

\item{opt.th.inline}{Upper bound to determine inline respondents at each iteration of 
optimization. A respondent is considered 'inline' if the distance between transformed
\code{ip2} and \code{ip1} goes below this threshold.}

\item{blend}{If \code{FALSE}, do not use blending procedure. The default is \code{TRUE}.}

\item{blend.th1}{If \code{blend==TRUE}, first threshold used in 'blending' procedure. If minimum difference 
between transformed \code{ip2} and \code{ip1} goes below this threshold, the 
transformed \code{ip2} is replaced with the closest \code{ip1}.}

\item{blend.th2}{If \code{blend==TRUE}, second threshold used in 'blending' procedure. If minimum difference 
between transformed \code{ip2} and \code{ip1} goes below this threshold (but above 
\code{blend.th1}), the transformed \code{ip2} is replaced with the value that 
intrapolates \code{ip1} that falls within this threshold.}
}
\value{
A list with the following elements along with specified argument values:
\itemize{
  \item \code{ip2_trans} or \code{ip1_trans}: Transformed matrix.
  \item \code{ip1}: Original \code{ip1} matrix.
  \item \code{ip2}: Original \code{ip2} matrix
}
}
\description{
Bridging Two Ideal Point Estimates with Linear Transformation Method
}
\author{
Tzu-Ping Liu \email{jamesliu0222@gmail.com}, Gento Kato \email{gento.badger@gmail.com}, and Sam Fuller \email{sjfuller@ucdavis.edu}.
}
