% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridge.olsmap.R
\name{bridge.olsmap}
\alias{bridge.olsmap}
\title{Bridging Two Ideal Point Estimates with OLS Mapping}
\usage{
bridge.olsmap(ip1, ip2, anchorrows.ip1, anchorrows.ip2)
}
\arguments{
\item{ip1}{Matrix or data.frame of 'reference' ideal points 
(i.e., \code{ip2} ideal points will be transformed and mapped to \code{ip1} space).
Rows are resepondents and columns are ideal point dimensions. 
The current code only allows 2 dimensions.}

\item{ip2}{Matrix or data.frame of ideal points to be transformed.}

\item{anchorrows.ip1}{Vector of row number of anchoring respondents in \code{ip1}.}

\item{anchorrows.ip2}{Vector of row number of anchoring respondents in \code{ip2}.
Must be the same length as \code{anchorrows.ip1}.}
}
\value{
A list with the following elements along with specified argument values:
\itemize{
  \item \code{ip1}: Original \code{ip1} matrix.
  \item \code{ip2_trans}: Transformed \code{ip2} matrix mapped on \code{ip1} space.
  \item \code{ip2_orig}: Original \code{ip2} matrix
}
}
\description{
Bridging Two Ideal Point Estimates with OLS Mapping
}
\references{
Shor, B.; Berry, C. & McCarty, N. (2010), 'A Bridge to Somewhere: Mapping State and Congressional Ideology on a Cross-institutional Common Space', Legislative Studies Quarterly 35(3), 417--448.
}
\author{
Tzu-Ping Liu \email{jamesliu0222@gmail.com}, Gento Kato \email{gento.badger@gmail.com}, and Sam Fuller \email{sjfuller@ucdavis.edu}.
}
