% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oocflex.R
\name{compute.nv}
\alias{compute.nv}
\title{Compute Normal Vector}
\usage{
compute.nv(xmat, resp, nv.method, ...)
}
\arguments{
\item{xmat}{Matrix of OC coordinates (i.e., predictors).}

\item{resp}{Response Variable (i.e., ordered choices).}

\item{nv.method}{Method of normal vector estimation. 
Following pre-defined methods are currently available:
\itemize{
  \item \code{"oprobit"}: Ordered probit regression (calls \code{\link[MASS]{polr}} function from \code{MASS} package)
  \item \code{"ologit"}: Ordered logistic regression (calls \code{\link[MASS]{polr}} function from \code{MASS} package)
  \item \code{"svm.reg"}: Support Vector Regression (calls \code{\link{nv.svm}} function)
  \item \code{"svm.class"}: Support Vector Classification with dichotomized response (calls \code{\link{nv.svm}} function)
  \item \code{"krls"}: Kernel Regularized Least Squares (calls \code{\link{nv.krls}} function). 
}
Alternatively, one can set custom made function that (at least) feeds in 
\code{xmat} and \code{resp} arguments and returns the vector of coefficients 
(the length must be equal to the number of columns of \code{xmat}).}

\item{...}{Additional arguments passed to the function assigned by \code{nv.method}.}
}
\value{
The vector of normal vector estimates.
}
\description{
Compute Normal Vector
}
\author{
Tzu-Ping Liu \email{jamesliu0222@gmail.com}, Gento Kato \email{gento.badger@gmail.com}, and Sam Fuller \email{sjfuller@ucdavis.edu}.
}
