% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipbridging.R
\name{ipbridging}
\alias{ipbridging}
\title{Bridging Ideal Point Estimates from Two Data Sets}
\usage{
ipbridging(d1, d2, bridge.method = "homography",
  anchors.method = "subsample", anchors.subsample.method = "random",
  anchors.subsample.pr = 0.1, anchors.subsample.wgt.d1 = NULL,
  anchors.subsample.wgt.d2 = NULL, anchors.selectrows.d1 = 1:100,
  anchors.selectrows.d2 = 1:100, anchors.newdata = NULL,
  input.type = "responses", ip.method = "ooc", ip.dims = 2,
  ip.polarity.d1 = c(1, 1), ip.polarity.d2 = c(1, 1),
  hg.opt.iter.n = 10000, hg.opt.sample.n = 30,
  hg.opt.th.inline = 0.5, hg.blend = TRUE, hg.blend.th1 = 0.05,
  hg.blend.th2 = 0.15, ...)
}
\arguments{
\item{d1}{Matrix or data.frame of 'reference' responses.
(i.e., In mapping methods, ideal points based on \code{d2} will be 
transformed and mapped to ideal point space based on \code{d1}).
Rows are resepondents and columns are questions/issues. 
Alternatively, if \code{input.type=="idealpoints"}, one can directly 
set matrix or data.frame of ideal points (rows are respondents and 
columns are ideal point dimensions).}

\item{d2}{Matrix or data.frame of responses (or ideal points) to be transformed.}

\item{bridge.method}{Method of bridging. Currently, following methods are aviailable:
\itemize{
  \item \code{"homography"} (default): Homography transformation method. 
  Based on anchor cases, this method provides non-parametric procedure to 
  find optimal transformation matrix to bridge ideal point estimates (calls \code{\link{bridge.homography}}).
  \item \code{"joint"}: Joint scaling method. Simple method to combine \code{d1} and 
  \code{d2} and jointly compute ideal points. The procedure described in Jessee (2016).
  \item \code{"modelmap"}: Use the model estimated on \code{d1} to map
  ideal point coordinates of \code{d2} respondents. This method does not require 
  anchor cases. Can be used only if \code{input.type=="responses"} and 
  \code{ip.method=="irtMCMC"}. See Jessee (2016).
  \item \code{"olsmap"}: OLS mapping method, Based on anchor cases, 
  use OLs regression to map \code{d2} ideal point coordinates on 
  \code{d1} ideal point space. The procedure described in Shor et al. (2010). 
  \item \code{"anchoredpooling"}: Use anchor cases to pool \code{d1} and \code{d2} 
  and jointly compute ideal points. Issues/questions in two datasets are assumed 
  to be independent from each other. See Shor et al. (2010) for more details.
}}

\item{anchors.method}{Method to select anchors to bridge ideal points. 
Following values are currently available. 
Ignored if \code{bridge.method} is \code{"joint"} or \code{"modelmap"}.
\itemize{
  \item \code{"subsample"} (default): Sample subset of \code{d1} and \code{d2} and 
  use them as anchors. Sampled anchors from another dataset are added to the 
  target dataset when computing ideal points on each dataset. 
  See \code{anchors.subsample.pr} and \code{anchors.subsample.method} for 
  sampling options.
  \item \code{"selectrows"}: Use this option if specific anchors are already 
  included in both \code{d1} and \code{d2}. Row numbers for anchors must be set 
  in \code{anchors.selectrows.d1} and \code{anchors.selectrows.d2}, respectively
  (Two vectors must have the same length and each element must corresponds to 
  the same anchor respondent). This is the only option for anchors 
  selection if \code{input.type=="idealpoints"}. 
  \item \code{"newdata"} Use new dataset that only contains anchors. The data.frame
  assigned in \code{anchors.newdata} are appended to both \code{d1} and \code{d2}. 
}
Note that \code{"subsample"} and \code{"newdata"} requires \code{d1} and \code{d2}
to have identical columns. This is not required for \code{"selectrows"}.}

\item{anchors.subsample.method}{Method to sample anchors. Following values 
are currently available: 
\itemize{
  \item \code{"random"} (default): randomly sample subset of both datasets and 
  use them as anchors. The sample size is determined by \code{anchors.subsample.pr}.
  \item \code{"random.d1"}: randomly sample subset of \code{d1} and 
  use them as anchors. The sample size is determined by \code{anchors.subsample.pr}.
  \item \code{"random.d2"}: randomly sample subset of \code{d2} and 
  use them as anchors. The sample size is determined by \code{anchors.subsample.pr}.
  \item \code{"selectrows"}: Use this option if specific anchors are already known (
  but not shared between datasets). Row numbers for anchors must be set 
  in \code{anchors.selectrows.d1} and \code{anchors.selectrows.d2}, respectively
  (Two vectors does not have to have the same length).
}}

\item{anchors.subsample.pr}{Used when  \code{anchors.method=="subsample"}. 
Proportion of dataset (i.e., \code{d1} and \code{d2}) 
to be sampled as anchor. The propotion is determined by the dataset of smaller size if 
\code{anchors.subsample.method=="random"}, specific dataset's size if 
\code{anchors.subsample.method} is \code{"random.d1"} or \code{"random.d2"}.}

\item{anchors.subsample.wgt.d1}{Optional weights passed to \code{prob} option of 
\code{\link[base]{sample}} function used in random sampling of anchors from \code{d1} when 
\code{anchors.subsample.method} is \code{"random"} or \code{"random.d1"}. If not \code{NULL}, 
it must be a vector of length equals to \code{nrow(d1)}.}

\item{anchors.subsample.wgt.d2}{Optional weights passed to \code{prob} option of 
\code{\link[base]{sample}} function used in random sampling of anchors from \code{d2} when 
\code{anchors.subsample.method} is \code{"random"} or \code{"random.d2"}. If not \code{NULL}, 
it must be a vector of length equals to \code{nrow(d1)}.}

\item{anchors.selectrows.d1}{Must be the vector of positive integers.
Specify the row numbers of anchors in \code{d1}. 
If \code{anchors.method=="selectrows"}, must be the same length and with perfect correspondence with \code{anchors.selectrows.d2}.
If \code{anchors.method=="subsample"} and \code{anchors.subsample.method=="selectrows"}, 
the above condition is not required.}

\item{anchors.selectrows.d2}{Must be the vector of positive integers.
Specify the row numbers of anchors in \code{d2}.}

\item{anchors.newdata}{Matrix or data.frame of anchors. Must have identical 
columns as \code{d1} and \code{d2}. Used if \code{anchors.method=="newdata"}.}

\item{input.type}{The input type of \code{d1} and \code{d2}. If \code{"responses"} (default), 
dataset values are responses to multiple issues or votes. 
If \code{"idealpoints"}, dataset values are ideal points.}

\item{ip.method}{Method of ideal point computation (used only when \code{input.type=="response"}. Following values are currently available.
\itemize{
  \item \code{"ooc"} (default): Ordered optimal classification. Calls \code{\link{oocflex}} function.
  \item \code{"blackbox"}: Blackbox scaling. Calls \code{\link[basicspace]{blackbox}} function from \code{blackbox} package (\code{minscale=10} by default).
  \item \code{"oc"}: Optimal classification. Calls \code{\link[oc]{oc}} function from \code{oc} package (\code{polarity=c(1,1)} by default). 
  \item \code{"wnominate"}: W-NOMINATE. Calls \code{\link[wnominate]{wnominate}} function from \code{wnominate} package  (\code{polarity=c(1,1)} by default).
  \item \code{"irtMCMC"}: IRT model via Markov chain Monte Carlo method. Calls \code{\link[pscl]{ideal}} function from \code{pscl} package.
}
For \code{ip.method=="ooc"}, the inputs \code{d1} and \code{d2} must be positive integer. 
For all other methods, the inputs must be a dummy variable indication 0=Nay and 1=Yea (and 9 for unit nonresponse).
All methods allow item non-response as NA.}

\item{ip.dims}{Number of dimension in ideal point computation. Must be a positive 
integer between 1 and 10. If \code{bridge.method=="homography"}, only 2 is accepted 
at this point.}

\item{ip.polarity.d1}{A vector specifying the row number of the \code{d1}(or pooled data)'s respondent(s) 
constrained to have a positive (i.e., right-wing or conservative) score on each dimension.
Used when \code{ip.method} is \code{"ooc"}, \code{"oc"}, or \code{"wnominate"}.}

\item{ip.polarity.d2}{A vector specifying the row number of the \code{d2}'s respondent(s) 
constrained to have a positive (i.e., right-wing or conservative) score on each dimension. 
Used when \code{ip.method} is \code{"ooc"}, \code{"oc"}, or \code{"wnominate"}.}

\item{hg.opt.iter.n}{Used if \code{bridge.method=="homography"}. 
Number of iteration in the optimization of transformation matrix.}

\item{hg.opt.sample.n}{Used if \code{bridge.method=="homography"}. 
Size of anchoring respondents to be sub-sampled at each iteration of 
optimization.}

\item{hg.opt.th.inline}{Used if \code{bridge.method=="homography"}. 
Upper bound to determine inline respondents at each iteration of 
optimization. A respondent is considered 'inline' if the distance between transformed
\code{ip2} and \code{ip1} goes below this threshold.}

\item{hg.blend}{Used if \code{bridge.method=="homography"}. 
If \code{FALSE}, do not use blending procedure in homograhy transformation. 
The default is \code{TRUE}.}

\item{hg.blend.th1}{Used if \code{bridge.method=="homography"}. 
First threshold used in 'blending' procedure. If minimum difference 
between transformed \code{ip2} and \code{ip1} goes below this threshold, the 
transformed \code{ip2} is replaced with the closest \code{ip1}.}

\item{hg.blend.th2}{Used if \code{bridge.method=="homography"}. Second threshold used in 'blending' procedure. If minimum difference 
between transformed \code{ip2} and \code{ip1} goes below this threshold (but above 
\code{blend.th1}), the transformed \code{ip2} is replaced with the value that 
intrapolates \code{ip1} that falls within this threshold.}

\item{...}{Additional arguments passed to the ideal point estimation function 
called in \code{ip.method}.}
}
\value{
A list with the following elements along with specified argument values
\itemize{
  \item \code{ip1}: Original \code{ip1} matrix.
  \item \code{ip2_trans}: Transformed \code{ip2} matrix mapped on \code{ip1} space.
  \item \code{ip2_orig}: Original \code{ip2} matrix
}
}
\description{
Toolbox function that Bridges ideal point estimates computed from 
two separate datasets.
}
\references{
\itemize{
  \item Jessee, S. (2016), '(How) Can We Estimate the Ideology of Citizens and Political Elites on the Same Scale?', American Journal of Political Science 60(4), 1108--1124.
  \item Shor, B.; Berry, C. & McCarty, N. (2010), 'A Bridge to Somewhere: Mapping State and Congressional Ideology on a Cross-institutional Common Space', Legislative Studies Quarterly 35(3), 417--448.
  
}
}
\seealso{
\code{\link{bridge.homography}}, \code{\link{oocflex}}, \code{\link[oc]{oc}}, 
\code{\link[wnominate]{wnominate}}, \code{\link[pscl]{ideal}}
}
\author{
Tzu-Ping Liu \email{jamesliu0222@gmail.com}, Gento Kato \email{gento.badger@gmail.com}, and Sam Fuller \email{sjfuller@ucdavis.edu}.
}
