% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipest.R
\name{ipest}
\alias{ipest}
\title{Estimating Ideal Points}
\usage{
ipest(
  d,
  method = "ooc",
  dims = 2,
  polarity = c(1, 1),
  irt.type = "binary",
  ...
)
}
\arguments{
\item{d}{A matrix with numeric values. With following conditions:
\itemize{
  \item Missing values (if any) must be set to \code{NA}.
  \item If variables are binary, values should be conisted from 0 and 1.
  \item If variables are ordered, values should be positive integers starting from 1.
}}

\item{method}{Method of ideal point computation (used only when \code{input.type=="response"}. Following values are currently available.
\itemize{
  \item \code{"ooc"} (default): Ordered optimal classification. Calls \code{\link{oocflex}} function.
  \item \code{"blackbox"}: Blackbox scaling. Calls \code{\link[basicspace]{blackbox}} function from \code{blackbox} package (\code{minscale=10} by default).
  \item \code{"oc"}: Optimal classification. Calls \code{\link[oc]{oc}} function from \code{oc} package (\code{polarity=c(1,1)} by default). 
  \item \code{"wnominate"}: W-NOMINATE. Calls \code{\link[wnominate]{wnominate}} function from \code{wnominate} package  (\code{polarity=c(1,1)} by default).
  \item \code{"irtMCMC"}: IRT model via Markov chain Monte Carlo method. 
  Calls \code{\link[pscl]{ideal}} function from \code{pscl} package if \code{irt.type=="binary"};
  Calls \code{\link[MCMCpack]{MCMCfactanal}} function from \code{MCMCpack} package if \code{irt.type=="ordered"} (still provisional).
  \item \code{"irtEM"}: IRT model via EM algorithm method (only one dimensional ideal points. This option is still provisional).
  Calls \code{\link[emIRT]{binIRT}} function from \code{emIRT} package if \code{irt.type=="binary"};
  Calls \code{\link[emIRT]{ordIRT}} function from \code{emIRT} package if \code{irt.type=="ordered"} (Only 3 category variable can be used).
}
For \code{method=="ooc"}, the inputs \code{d1} and \code{d2} must be positive integer. 
For all other methods, the inputs must be a dummy variable indication 0=Nay and 1=Yea (and 9 for unit nonresponse).
All methods allow item non-response as NA.}

\item{dims}{Number of dimension in ideal point computation. Must be a positive 
integer between 1 and 10. If \code{bridge.method=="homography"}, only 2 is accepted 
at this point.}

\item{polarity}{A vector specifying the row number of the \code{d1}(or pooled data)'s respondent(s) 
constrained to have a positive (i.e., right-wing or conservative) score on each dimension.
Used when \code{method} is \code{"ooc"}, \code{"oc"}, or \code{"wnominate"}.}

\item{irt.type}{Used if \code{method} is \code{"irtMCMC"} or \code{"irtEM"}. 
If \code{"binary"} the binary IRT will be estimated. If \code{"ordered"}, 
the ordered IRT will be estimated (the ordered option is provisional).}

\item{...}{Additional arguments passed to the ideal point estimation function 
called in \code{method}.}
}
\value{
A list with the following elements
\itemize{
  \item \code{ip}: Matrix of idealpoints.
  \item \code{info}: Outputs including other information.
}
}
\description{
Toolbox function that estimates ideal points using different methods.
}
\references{
\itemize{
  \item Jessee, S. (2016), '(How) Can We Estimate the Ideology of Citizens and Political Elites on the Same Scale?', American Journal of Political Science 60(4), 1108--1124.
  \item Shor, B.; Berry, C. & McCarty, N. (2010), 'A Bridge to Somewhere: Mapping State and Congressional Ideology on a Cross-institutional Common Space', Legislative Studies Quarterly 35(3), 417--448.
  
}
}
\seealso{
\code{\link{ipbridging}}, \code{\link{oocflex}}, \code{\link[oc]{oc}}, 
\code{\link[wnominate]{wnominate}}, \code{\link[pscl]{ideal}}
}
\author{
Tzu-Ping Liu \email{jamesliu0222@gmail.com}, Gento Kato \email{gento.badger@gmail.com}, and Sam Fuller \email{sjfuller@ucdavis.edu}.
}
