% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nv.krls.R
\name{nv.krls}
\alias{nv.krls}
\title{Estimating Normal Vector through Kernel Regularized Least Squares}
\usage{
nv.krls(xmat, resp, sample.size = 300, ...)
}
\arguments{
\item{xmat}{Matrix of OC coordinates (i.e., predictors).}

\item{resp}{Response Variable (i.e., ordered choices).}

\item{sample.size}{Sample size for data used for the estimation of KRLS.}

\item{...}{Additional arguments passed to \code{\link[KRLS]{krls}}.}
}
\value{
A vector of coefficients.
}
\description{
Estimating Normal Vector through Kernel Regularized Least Squares
}
\references{
Jeremy Ferwerda, Jens Hainmueller, Chad J. Hazlett (2017). Kernel-Based Regularized Least Squares in R (KRLS) and Stata (krls). Journal of Statistical Software, 79(3), 1-26. doi:10.18637/jss.v079.i03
}
\seealso{
\code{\link[KRLS]{krls}}
}
\author{
Tzu-Ping Liu \email{jamesliu0222@gmail.com}, Gento Kato \email{gento.badger@gmail.com}, and Sam Fuller \email{sjfuller@ucdavis.edu}.
This is a modified version of the code included in \code{\link[ooc]{ooc}} function.
}
