% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oocflex.R
\name{oocflex}
\alias{oocflex}
\title{Ordered Optimal Classification (Flexible Version)}
\usage{
oocflex(
  votemat,
  dims = 2,
  minvotes = 10,
  lop = 0.001,
  polarity = c(1, 1),
  iter = 25,
  nv.method = "svm.reg",
  binary.method = "dominance",
  random.seed = NULL,
  ...
)
}
\arguments{
\item{votemat}{A matrix of ordinal choice data, must be consecutive 
integers starting with 1. Binary choice should be coded so that 
1 = Support, 2 = Oppose. Missing data must be coded as NA.}

\item{dims}{Number of dimensions to estimate.}

\item{minvotes}{Minimum number of votes required for a respondent to be 
included in the analysis.}

\item{lop}{A proportion between 0 and 1, the cut-off used for excluding 
lopsided votes.}

\item{polarity}{A vector specifying the row number of the respondent(s) 
constrained to have a positive (i.e., right-wing or conservative) score 
on each dimension.}

\item{iter}{Number of iterations of the modified Optimal Classification 
algorithm.}

\item{nv.method}{The method used to compute normal vectors at each step 
of the iteration. Current choices include: 
\code{"oprobit"} (Ordered probit regression), 
\code{"ologit"} (Ordered logistic regression), 
\code{"svm.reg"} (SVM: regression), 
\code{"svm.class"} (SVM: classification), 
and \code{"krls"} (Kernel regularized least squares). 
One can also assign a custom-made function to compute normal vectors.
See more details in \code{\link{compute.nv}}.}

\item{binary.method}{The method used to transform ordinal responses 
to binary ones. Following options are currently available:
\itemize{
  \item \code{"dominance"} (default): Transformed to the matrix with binary choices organized in a dominance pattern.
  \item \code{"even"}: Transformed to the matrix with binary choices recoded to be as evenly-balanced as possible.
}}

\item{random.seed}{If not \code{NULL}, scalar indicates the random seed for reproducibility.}

\item{...}{Additional arguments passed to the function assigned 
by \code{nv.method} (See more details in \code{\link{compute.nv}}).}
}
\value{
An \code{oocflexObject} with the following elements
\itemize{
  \item \code{respondents}: A matrix containing the respondent estimates.
  \item \code{issues}: A matrix containing the estimates for the (c-1) categories of each issue.
  \item \code{issues.unique}: A matrix containing only the estimates for the unique issues.
  \item \code{fits}: The percentage of all binary choices correctly predicted, 
  the Aggregate Proportional Reduction in Error (APRE) statistic for the binary choices, 
  the percentage of all ordinal choices correctly predicted, 
  and the Aggregate Proportional Reduction in Error (APRE) statistic for 
  the ordinal choices.
  \item \code{OC.result.binary}: Standard (binary) OC results 
  for the choice matrix arranged in a binary dominance pattern.
  \item \code{errorcount}: Error counts at each iteration of the Optimal Classification algorithm.
  \item \code{votemat.binary}: The matrix of ordinal choices recoded into binary categories 
  that are as balanced as possible.
}
}
\description{
Performs Ordered Optimal Classification (OOC) with flexible strategies of estimation. OOC is an extension of Poole's (2000) nonparametric unfolding procedure for the analysis of ordinal choice data (e.g., “Strongly Agree,” “Somewhat Agree,” “Somewhat Disagree,” “Strongly Disagree”).
}
\references{
\itemize{
  \item Hare, C., Liu, T., & Lupton, R. N. 2018. "What Ordered Optimal Classification Reveals about Ideological Structure, Cleavages, and Polarization in the American Mass Public", Public Choice, 176, 57-78. 
  \item Armstrong, David A., Ryan Bakker, Royce Carroll, Christopher Hare, Keith T. Poole, and Howard Rosenthal. 2014. Analyzing Spatial Models of Choice and Judgment with R. Boca Raton, FL: CRC Press.
  \item Poole, Keith T. 2000. "Nonparametric unfolding of BInary Choice Data." Political Analysis 8(e): 211-237.       
}
}
\seealso{
\code{\link{compute.nv}}, \code{\link{nv.svm}}, and \code{\link{nv.krls}}
}
\author{
Tzu-Ping Liu \email{jamesliu0222@gmail.com}, Gento Kato \email{gento.badger@gmail.com}, and Sam Fuller \email{sjfuller@ucdavis.edu}.
This code is the modified version of the \code{\link[ooc]{ooc}} 
function written by Christopher Hare and Keith T. Poole.
}
