% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setanchors.R
\name{setanchors}
\alias{setanchors}
\title{Setting Anchors to Combine Two Datasets}
\usage{
setanchors(
  d1,
  d2,
  anchors.method = "subsample",
  anchors.subsample.method = "random",
  anchors.subsample.pr = 0.1,
  anchors.subsample.wgt.d1 = NULL,
  anchors.subsample.wgt.d2 = NULL,
  anchors.selectrows.d1 = 1:100,
  anchors.selectrows.d2 = 1:100,
  anchors.selectrows.data = NULL,
  anchors.newdata = NULL
)
}
\arguments{
\item{d1}{Matrix or data.frame of 'reference' responses.
(i.e., In mapping methods, ideal points based on \code{d2} will be 
transformed and mapped to ideal point space based on \code{d1}).
Rows are resepondents and columns are questions/issues.}

\item{d2}{Matrix or data.frame of responses to be transformed.}

\item{anchors.method}{Method to select anchors to bridge ideal points. 
Following values are currently available. 
Ignored if \code{bridge.method} is \code{"joint"} or \code{"modelmap"}.
\itemize{
  \item \code{"subsample"} (default): Sample subset of \code{d1} and \code{d2} and 
  use them as anchors. Sampled anchors from another dataset are added to the 
  target dataset when computing ideal points on each dataset. 
  See \code{anchors.subsample.pr} and \code{anchors.subsample.method} for 
  sampling options.
  \item \code{"selectrows"}: Use this option if specific anchors are already 
  included in both \code{d1} and \code{d2}. Row numbers for anchors must be set 
  in \code{anchors.selectrows.d1} and \code{anchors.selectrows.d2}, respectively
  (Two vectors must have the same length and each element must corresponds to 
  the same anchor respondent). This is the only option for anchors 
  selection if \code{input.type=="idealpoints"}. 
  \item \code{"newdata"} Use new dataset that only contains anchors. The data.frame
  assigned in \code{anchors.newdata} are appended to both \code{d1} and \code{d2}. 
}
Note that \code{"subsample"} and \code{"newdata"} requires \code{d1} and \code{d2}
to have identical columns. This is not required for \code{"selectrows"}.}

\item{anchors.subsample.method}{Method to sample anchors. Following values 
are currently available: 
\itemize{
  \item \code{"random"} (default): randomly sample subset of both datasets and 
  use them as anchors. The sample size is determined by \code{anchors.subsample.pr}.
  \item \code{"random.d1"}: randomly sample subset of \code{d1} and 
  use them as anchors. The sample size is determined by \code{anchors.subsample.pr}.
  \item \code{"random.d2"}: randomly sample subset of \code{d2} and 
  use them as anchors. The sample size is determined by \code{anchors.subsample.pr}.
  \item \code{"selectrows"}: Use this option if specific anchors are already known (
  but not shared between datasets). Row numbers for anchors must be set 
  in \code{anchors.selectrows.d1} and \code{anchors.selectrows.d2}, respectively
  (Two vectors does not have to have the same length).
}}

\item{anchors.subsample.pr}{Used when  \code{anchors.method=="subsample"}. 
Proportion of dataset (i.e., \code{d1} and \code{d2}) 
to be sampled as anchor. The propotion is determined by the dataset of smaller size if 
\code{anchors.subsample.method=="random"}, specific dataset's size if 
\code{anchors.subsample.method} is \code{"random.d1"} or \code{"random.d2"}.}

\item{anchors.subsample.wgt.d1}{Optional weights passed to \code{prob} option of 
\code{\link[base]{sample}} function used in random sampling of anchors from \code{d1} when 
\code{anchors.subsample.method} is \code{"random"} or \code{"random.d1"}. If not \code{NULL}, 
it must be a vector of length equals to \code{nrow(d1)}.}

\item{anchors.subsample.wgt.d2}{Optional weights passed to \code{prob} option of 
\code{\link[base]{sample}} function used in random sampling of anchors from \code{d2} when 
\code{anchors.subsample.method} is \code{"random"} or \code{"random.d2"}. If not \code{NULL}, 
it must be a vector of length equals to \code{nrow(d1)}.}

\item{anchors.selectrows.d1}{Must be the vector of positive integers.
Specify the row numbers of anchors in \code{d1}. 
If \code{anchors.method=="selectrows"}, must be the same length and with perfect correspondence with \code{anchors.selectrows.d2}.
If \code{anchors.method=="subsample"} and \code{anchors.subsample.method=="selectrows"}, 
the above condition is not required.}

\item{anchors.selectrows.d2}{Must be the vector of positive integers.
Specify the row numbers of anchors in \code{d2}.}

\item{anchors.selectrows.data}{Required if \code{anchors.method=="selectrows"}. 
Specify the dataset where anchors are originally found. 
Must be the same length as \code{anchors.selectrows.d1}. Allowed values are 1 (=\code{d1}), 
2 (=\code{d2}) and 3 (=new data).}

\item{anchors.newdata}{Matrix or data.frame of anchors. Must have identical 
columns as \code{d1} and \code{d2}. Used if \code{anchors.method=="newdata"}.}
}
\value{
A list with the following elements:
\itemize{
  \item \code{bridge.data}: data.frame to store the ideal point estimates in the next method.
  \item \code{d.ip1}: \code{d1} with anchors.
  \item \code{d.ip2}: \code{d2} with anchors.
  \item \code{anchorrows.ip1}: Anchor locations in \code{d.ip1}.
  \item \code{anchorrows.ip2}: Anchor locations in \code{d.ip2}.
  \item \code{anchorrows.ip1.data}: Original dataset for anchors identified in \code{anchorrows.ip1}.
  \item \code{anchorrows.ip2.data}: Original dataset for anchors identified in \code{anchorrows.ip2}.
}
}
\description{
Toolbox function that sets anchors to bridge two datasets.
}
\seealso{
\code{\link{ipbridging}}, \code{\link{ipest}}
}
\author{
Tzu-Ping Liu \email{jamesliu0222@gmail.com}, Gento Kato \email{gento.badger@gmail.com}, and Sam Fuller \email{sjfuller@ucdavis.edu}.
}
