% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.oocflexObject}
\alias{summary.oocflexObject}
\title{Ordered Optimal Classification Summary}
\usage{
\method{summary}{oocflexObject}(object, verbose = FALSE, ...)
}
\arguments{
\item{object}{an \code{oocflexObject} output object.}

\item{verbose}{logical, includes all ideal points if TRUE, 
otherwise only returns the first 10 respondents.}

\item{...}{Other arguments. Do nothing and are not passed to any functions.}
}
\value{
A summary of a \code{oocflexObject}.
}
\description{
\code{summary.oocflexObject} reads an Ordered Optimal 
Classification object exported by \code{\link{oocflex}} function and
prints a summary.
}
\seealso{
\code{\link{oocflex}}
}
\author{
Tzu-Ping Liu \email{jamesliu0222@gmail.com}, Gento Kato \email{gento.badger@gmail.com}, and Sam Fuller \email{sjfuller@ucdavis.edu}.
This code is the modified version of the \code{\link[oc]{summary.OCobject}} 
method written by Keith Poole, Jeffrey Lewis, James Lo, and Royce Carroll.
}
