% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nv.svm.R
\name{nv.svm}
\alias{nv.svm}
\title{Estimating Normal Vector through Support Vector Machine}
\usage{
nv.svm(xmat, resp, tune.param = "heuristics", tune.gpu = FALSE,
  param.heuristics.k = 3, ...)
}
\arguments{
\item{xmat}{Matrix of OC coordinates (i.e., predictors).}

\item{resp}{Response Variable (i.e., ordered choices).}

\item{tune.param}{Method to determine the parameters. 
Following options are currently available:
\itemize{
  \item \code{NULL}: Use \code{\link[e1071]{svm}} with the default 
  or manually set parameter values. To manually set parameters.
  Check the available parameters (and their default values) in 
  \code{\link[e1071]{svm}} function in \code{e1071} package.
  \item \code{"heuristics"} (default): If \code{resp} is numeric, 
  use heuristic based method described in Cherkassky & Ma (2004) to 
  determine optimal parameters for C (cost) and epsilon.
  If \code{resp} is a factor, the same behavior as \code{NULL}. 
  \item A named list of parameter vectors spanning the sampling space. 
  If list is assigned here, the method use \code{\link[e1071]{best.tune}} function
  to obtain the best perfomed model in the given parameter ranges. 
}}

\item{tune.gpu}{If \code{TRUE} and tuning is conducted, 
the method replaces \code{\link[e1071]{best.tune}} function with 
the \code{best.tune} function in \code{Rgtsvm} package, which contains
GPU accelerate method of \code{svm} function. Install \code{Rgtsvm} manually. 
You need to have Linux computer with NVIDIA GPU for the installation.
Check \url{https://github.com/Danko-Lab/Rgtsvm} for more details.}

\item{param.heuristics.k}{If \code{tune.param=="heuristics"}, the method uses 
k nearest neighbors regression to estimate the noise variance in response variable.
This argument sets k for this method. The default is 3. Cherkassky & Ma (2004) 
recommends somewhere between 2 to 6.}

\item{...}{Additional arguments passed to \code{\link[e1071]{svm}} 
or \code{\link[e1071]{best.tune}}.}
}
\value{
A vector of coefficients.
}
\description{
Estimating Normal Vector through Support Vector Machine
}
\references{
\itemize{
  \item Chang, C. & C. Lin, LIBSVM: A Library for Support Vector Machines, \url{http://www.csie.ntu.edu.tw/~cjlin/libsvm}.
  \item Cherkassky, V. & Y. Ma, 2004, "Practical Selection of SVM Parameters and Noise Estimation for SVM Regression", Neural Networks, 17, 113 - 126.
  \item Cotter, A., N. Srebro, & J. Kreshet, 2011, "A GPU-Tailored Approach for Training Kernelized SVMs", 17th ACM SIGKDD Conference on Knowledge Discovery and Data Mining.
}
}
\seealso{
\code{\link[e1071]{svm}}, \code{\link[e1071]{best.tune}}
}
\author{
Tzu-Ping Liu \email{jamesliu0222@gmail.com}, Gento Kato \email{gento.badger@gmail.com}, and Sam Fuller \email{sjfuller@ucdavis.edu}.
This is a modified version of the code included in \code{\link[ooc]{ooc}} function.
}
